Refactored to a single TLS package
