+++ src/main/java/com/xjeffrose/xio2/TLS/SelfSignedCertGenerator.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.xjeffrose.xio2.http.server.TLS;
+package com.xjeffrose.xio2.TLS;
 
 import com.xjeffrose.log.Log;
 import java.math.BigInteger;
