+++ src/main/java/com/xjeffrose/xio2/TLS/XioHostnameVerifier.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.xjeffrose.xio2.http.client.TLS;
+package com.xjeffrose.xio2.TLS;
 
 import java.security.cert.Certificate;
 import java.security.cert.CertificateParsingException;
