+++ src/main/java/com/xjeffrose/xio2/TLS/xioCertificate.java
@@ -15,7 +15,7 @@
  *
  */
 
-package com.xjeffrose.xio2.http.server.TLS;
+package com.xjeffrose.xio2.TLS;
 
 import com.xjeffrose.log.Log;
 import java.security.PrivateKey;
