+++ src/main/java/com/xjeffrose/xio2/http/client/Client.java
@@ -23,7 +23,7 @@ import com.xjeffrose.xio2.http.HttpResponseParser;
 import com.xjeffrose.xio2.http.client.LoadBalancerStrategies.LoadBalancingStrategy;
 import com.xjeffrose.xio2.http.client.LoadBalancerStrategies.NullLoadBalancer;
 import com.xjeffrose.xio2.http.client.LoadBalancerStrategies.RoundRobinLoadBalancer;
-import com.xjeffrose.xio2.http.client.TLS.TLS;
+import com.xjeffrose.xio2.TLS.TLS;
 import com.xjeffrose.xio2.ChannelContext;
 import java.io.IOException;
 import java.net.InetSocketAddress;
@@ -138,7 +138,7 @@ public class Client {
   private HttpResponse execute(SocketChannel channel) {
     if (_tls) {
       if (checkConnect(channel)) {
-        if (tls.execute()) {
+        if (tls.doHandshake()) {
           if (write(channel)) {
             return read(channel);
           }
