+++ src/test/java/com/xjeffrose/xio2/TLS/KeyStoreFactoryTest.java
@@ -1,4 +1,4 @@
-package com.xjeffrose.xio2.http.server.TLS;
+package com.xjeffrose.xio2.TLS;
 
 import java.security.KeyStore;
 import org.junit.Test;
