--- src/test/java/com/xjeffrose/xio2/http/server/TLS/TLSTest.java
@@ -1,66 +1,0 @@
-/*
- * Copyright (C) 2015 Jeff Rose
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.xjeffrose.xio2.http.server.TLS;
-
-import org.junit.After;
-import org.junit.Before;
-import org.junit.Test;
-
-public class TLSTest {
-
-  @Before
-  public void setUp() throws Exception {
-
-  }
-
-  @After
-  public void tearDown() throws Exception {
-
-  }
-
-  @Test
-  public void testEncrypt() throws Exception {
-
-  }
-
-  @Test
-  public void testDecrypt() throws Exception {
-
-  }
-
-  @Test
-  public void testClose() throws Exception {
-
-  }
-
-  @Test
-  public void testTerminate() throws Exception {
-
-  }
-
-  @Test
-  public void testEngineBuilder() throws Exception {
-//    SSLEngine engine = TLS.engineBuilder();
-//    assertTrue(engine != null);
-//    assertEquals(engine.getEnabledProtocols()[0], "TLSv1.2");
-//    assertEquals(engine.getEnabledCipherSuites()[0], "TLS_ECDHE_RSA_WITH_AES_128_GCM_SHA256");
-  }
-
-  @Test
-  public void testEncryption() throws Exception {
-  }
-
-}
